+++ tpe2-g7/api/src/main/java/ar/edu/itba/pod/MapReduce/reducers/Query1ReducerFactory.java
@@ -5,9 +5,9 @@ import com.hazelcast.mapreduce.Reducer;
 import com.hazelcast.mapreduce.ReducerFactory;
 
 //EJEMPLO EXTRAIDO DE LA TEORICA
-public class Query1ReducerFactory implements ReducerFactory<Pair<Integer>, Long, Long> {
+public class Query1ReducerFactory implements ReducerFactory<Pair<Long>, Long, Long> {
     @Override
-    public Reducer<Long, Long> newReducer(Pair<Integer> integerPair) {
+    public Reducer<Long, Long> newReducer(Pair<Long> integerPair) {
         return new Query1Reducer();
     }
